refactor(rye-app): extract main-thread launcher + DOM forwarders + postMessage helpers to wasm/main_launcher.rs (was inlined in worker.rs); re-export launch_on_click at wasm module root
